requirements.txt
torch
torchvision
Pillow
aiogram
python-decouple